# cmd/agent/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: agent-sandbox-agent
  version: 1.0.0
  description: RESTful API exposed by the sandbox agent sidecar process.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/files:
    get:
      tags: [files]
      summary: List directory entries
      operationId: listFiles
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
            example: /workspace/myproject
          description: Absolute path of directory to list
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileListResponse'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [files]
      summary: Create file or directory
      operationId: createFile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileEntry'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [files]
      summary: Delete file or directory
      operationId: deleteFile
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
          description: Absolute path of file or directory to delete
        - name: recursive
          in: query
          schema:
            type: boolean
            default: false
          description: Delete directory recursively
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/files/content:
    get:
      tags: [files]
      summary: Read file content
      operationId: readFileContent
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
          description: Absolute path of file to read
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileContentResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Error'
    put:
      tags: [files]
      summary: Write file content
      operationId: writeFileContent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileContentRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileEntry'
        default:
          $ref: '#/components/responses/Error'

  # ── Git ───────────────────────────────────────────────────────────────────
  /api/v1/git/status:
    get:
      tags: [git]
      summary: Get repository status
      operationId: getGitStatus
      parameters:
        - name: repoPath
          in: query
          required: true
          schema:
            type: string
          description: Absolute path of the git repository root
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GitStatusResponse'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/git/branches:
    get:
      tags: [git]
      summary: List branches
      operationId: listGitBranches
      parameters:
        - name: repoPath
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GitBranchListResponse'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/git/diff:
    get:
      tags: [git]
      summary: Get file diff
      operationId: getGitDiff
      parameters:
        - name: repoPath
          in: query
          required: true
          schema:
            type: string
        - name: filePath
          in: query
          schema:
            type: string
          description: Relative file path within repo; omit for full diff
        - name: staged
          in: query
          schema:
            type: boolean
            default: false
          description: Show staged (cached) diff
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GitDiffResponse'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/git/commit:
    post:
      tags: [git]
      summary: Commit changes
      operationId: gitCommit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GitCommitRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GitCommitResponse'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/git/checkout:
    post:
      tags: [git]
      summary: Checkout branch
      operationId: gitCheckout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GitCheckoutRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GitStatusResponse'
        default:
          $ref: '#/components/responses/Error'

  # ── Terminals ─────────────────────────────────────────────────────────────
...
components:
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 404
        message:
          type: string
          example: resource not found

    # ── Projects ──────────────────────────────────────────────────────────────
...
    FileEntry:
      type: object
      required: [name, path, type]
      properties:
        name:
          type: string
        path:
          type: string
        type:
          type: string
          enum: [file, directory, symlink]
        size:
          type: integer
          format: int64
          description: File size in bytes (0 for directories)
        modifiedAt:
          type: string
          format: date-time
        permissions:
          type: string
          example: "0644"
    FileListResponse:
      type: object
      required: [path, entries]
      properties:
        path:
          type: string
        entries:
          type: array
          items:
            $ref: '#/components/schemas/FileEntry'
    FileCreateRequest:
      type: object
      required: [path, type]
      properties:
        path:
          type: string
          description: Absolute path of the new file or directory
        type:
          type: string
          enum: [file, directory]
        content:
          type: string
          description: Initial content (base64-encoded) for file type
        permissions:
          type: string
          example: "0644"
    FileContentRequest:
      type: object
      required: [path, content]
      properties:
        path:
          type: string
          description: Absolute path of the file to write
        content:
          type: string
          description: File content (base64-encoded)
        encoding:
          type: string
          enum: [base64, utf8]
          default: utf8
    FileContentResponse:
      type: object
      required: [path, content, encoding]
      properties:
        path:
          type: string
        content:
          type: string
        encoding:
          type: string
          enum: [base64, utf8]
        size:
          type: integer
          format: int64

    # ── Git ───────────────────────────────────────────────────────────────────
    GitFileStatus:
      type: object
      required: [path, status]
      properties:
        path:
          type: string
        status:
          type: string
          enum: [modified, added, deleted, renamed, untracked, conflicted]
        oldPath:
          type: string
          description: Original path when status is renamed
    GitStatusResponse:
      type: object
      required: [repoPath, branch, clean]
      properties:
        repoPath:
          type: string
        branch:
          type: string
        clean:
          type: boolean
        ahead:
          type: integer
          description: Commits ahead of remote
        behind:
          type: integer
          description: Commits behind remote
        staged:
          type: array
          items:
            $ref: '#/components/schemas/GitFileStatus'
        unstaged:
          type: array
          items:
            $ref: '#/components/schemas/GitFileStatus'
        untracked:
          type: array
          items:
            type: string
    GitBranch:
      type: object
      required: [name, current]
      properties:
        name:
          type: string
        current:
          type: boolean
        remote:
          type: string
        lastCommit:
          type: string
    GitBranchListResponse:
      type: object
      required: [branches]
      properties:
        branches:
          type: array
          items:
            $ref: '#/components/schemas/GitBranch'
    GitDiffResponse:
      type: object
      required: [repoPath, diff]
      properties:
        repoPath:
          type: string
        filePath:
          type: string
        diff:
          type: string
          description: Unified diff output
    GitCommitRequest:
      type: object
      required: [repoPath, message]
      properties:
        repoPath:
          type: string
        message:
          type: string
        paths:
          type: array
          items:
            type: string
          description: Specific files to stage; empty means stage all changes
        authorName:
          type: string
        authorEmail:
          type: string
    GitCommitResponse:
      type: object
      required: [commitHash, message, author, committedAt]
      properties:
        commitHash:
          type: string
        message:
          type: string
        author:
          type: string
        committedAt:
          type: string
          format: date-time
    GitCheckoutRequest:
      type: object
      required: [repoPath, branch]
      properties:
        repoPath:
          type: string
        branch:
          type: string
        createNew:
          type: boolean
          default: false
          description: Create a new branch if it does not exist

    # ── Terminals ─────────────────────────────────────────────────────────────
...